requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.2
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
PyJWT==2.10.1
pymongo==4.15.5
python-dotenv==1.2.1
Werkzeug==3.1.4
